Wisdom of Sea and Sky
The new set introduces a new type of booster -- regular plus one card.
The plus one cards are unique to this type of booster, so supporting
them in the simulation is not so straight forward.

See expansions/format.md for the updated design and rationale.

TODO: add new unit tests to verify the new behaviour is correct (I
believe it is).