Mega Rising
The main change here is that the shiny (rainbow) cards are only
available as the 6th card in regular+1 booster, tho the actual
probability of pulling one is about the same as previous expansions.

This also means that rainbow cards no longer appear in rare boosters. To
account for this, I added a new 'common' field to make explict whether a
rarity is 'common' (1 to 4 diamonds), rather than assuming anything
which doesn't appear in rare boosters is common.

For backwards compatibility, if not set 'common' defaults to not rare.
This means I don't have to retroactively update all the expansion data.

Mega Rising is the harded expansion to fully complete to date.
Deluxe ex still has the edge for completing common cards due to all the
mirror holos.